AWSSecurityAnalyticsBootstrap/cdk/requirements.txt
-e .
boto3
aws-cdk-lib>=2.2.0
constructs>=10.0.0